+++ src/main/java/com/atlauncher/gui/tabs/InstancesTab.java
@@ -43,6 +43,7 @@ import com.atlauncher.gui.card.NilCard;
 import com.atlauncher.gui.dialogs.AddCurseForgePackDialog;
 import com.atlauncher.gui.dialogs.AddFTBPackDialog;
 import com.atlauncher.gui.dialogs.ImportInstanceDialog;
+import com.atlauncher.managers.ConfigManager;
 import com.atlauncher.managers.InstanceManager;
 import com.atlauncher.network.Analytics;
 
@@ -112,10 +113,17 @@ public class InstancesTab extends JPanel implements Tab, RelocalizationListener 
 
         topPanel.add(importButton);
         topPanel.add(Box.createHorizontalStrut(5));
-        topPanel.add(addCurseForgePackButton);
-        topPanel.add(Box.createHorizontalStrut(5));
-        topPanel.add(addFTBPackButton);
-        topPanel.add(Box.createHorizontalGlue());
+
+        if (ConfigManager.getConfigItem("platforms.curseforge.modpacksEnabled", true) == true) {
+            topPanel.add(addCurseForgePackButton);
+            topPanel.add(Box.createHorizontalStrut(5));
+        }
+
+        if (ConfigManager.getConfigItem("platforms.modpacksch.modpacksEnabled", true) == true) {
+            topPanel.add(addFTBPackButton);
+            topPanel.add(Box.createHorizontalGlue());
+        }
+
         topPanel.add(searchField);
         topPanel.add(Box.createHorizontalStrut(5));
         topPanel.add(searchButton);