+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This changelog only contains the changes that are unreleased. For changes for in
 ## 3.4.6.1
 
 ### New Features
+- Add in launcher config to be able to toggle and disable things without a launcher update
 
 ### Fixes
 - Update the LaunchServer.bat script to make it "simpler" to change Java path